Add files via upload
Write a phonebook with graphical user interface and sql commands as database in python that you can see some modules which we import them in our program for example Tkinter, Sqlite3, and Tabulate, PrettyTable, MessageBox, Image and Image TK from PIL. I've done it very creatively in the best form that user can understand my program and work with it easily & correctly. This project is one of my best projects in coding with python as back-end programming language. And also I worked this project in my Python course with my dear teacher Khosrow Hafezi Motlagh in Radman Iran Institute.